+++ go-client-sdk/sdk.go
@@ -85,8 +85,8 @@ func WithClient(client HTTPClient) SDKOption {
 func New(opts ...SDKOption) *SDK {
 	sdk := &SDK{
 		_language:   "go",
-		_sdkVersion: "0.3.0",
-		_genVersion: "2.26.0",
+		_sdkVersion: "0.3.1",
+		_genVersion: "2.26.1",
 	}
 	for _, opt := range opts {
 		opt(sdk)