+++ typescript-client-sdk/src/sdk/sdk.ts
@@ -47,8 +47,8 @@ export class SDK {
   public _securityClient: AxiosInstance;
   public _serverURL: string;
   private _language = "typescript";
-  private _sdkVersion = "0.2.0";
-  private _genVersion = "2.26.0";
+  private _sdkVersion = "0.1.2";
+  private _genVersion = "2.26.1";
   private _globals: any;
 
   constructor(props?: SDKProps) {