+++ RELEASES.md
@@ -93,3 +93,13 @@ Based on:
 - [Go v0.3.0] https://github.com/speakeasy-sdks/brease/releases/tag/go-client-sdk/v0.3.0 - go-client-sdk
 - [PyPI v1.2.0] https://pypi.org/project/brease-sdk/1.2.0 - python-client-sdk
 - [NPM v0.2.0] https://www.npmjs.com/package/brease-sdk/v/0.2.0 - typescript-client-sdk
+
+## 2023-05-06 00:37:30
+### Changes
+Based on:
+- OpenAPI Doc 0.1.0 https://prod.speakeasyapi.dev/v1/apis/brease/version/0.1/schema/download
+- Speakeasy CLI 1.29.1 (2.26.1) https://github.com/speakeasy-api/speakeasy
+### Releases
+- [PyPI v1.1.1] https://pypi.org/project/brease-sdk/1.1.1 - python-client-sdk
+- [NPM v0.1.2] https://www.npmjs.com/package/brease-sdk/v/0.1.2 - typescript-client-sdk
+- [Go v0.3.1] https://github.com/speakeasy-sdks/brease/releases/tag/go-client-sdk/v0.3.1 - go-client-sdk